Most generation options work Added unit tests with HTML outputs to see the grid Added a better word counter

+++ CrosswordGenerator/Extensions/IntExtensions.cs
@@ -12,7 +12,7 @@
             {
                 modulo = (dividend - 1) % 26;
                 columnLabel = Convert.ToChar(65 + modulo).ToString() + columnLabel;
-                dividend = (int)((dividend - modulo) / 26);
+                dividend = (dividend - modulo) / 26;
             }
 
             return columnLabel;
